transformlivedata/requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
certifi==2026.1.4
cffi==2.0.0
dotenv==0.9.9
minio==7.2.20
psycopg2-binary==2.9.11
pycparser==2.23
pycryptodome==3.23.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
six==1.17.0
typing_extensions==4.15.0
urllib3==2.6.3